requirements.txt
streamlit==1.26.0
python-dotenv
requests
